implementa valida entrada, corrige valor entrada de Double para BigDecimal

+++ src/main/java/com/rlti/autoescola/matricula/application/api/MatriculaAlteracaoRequest.java
@@ -3,10 +3,12 @@ package com.rlti.autoescola.matricula.application.api;
 import com.rlti.autoescola.matricula.domain.TipoPagamento;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 public class MatriculaAlteracaoRequest {
     TipoPagamento tipoPagamento;
-    Double valorEntrada;
+    BigDecimal valorEntrada;
     int desconto;
     int quantidadeParcelas;
     String observacao;
