+++ src/main/java/com/rlti/autoescola/matricula/application/api/MatriculaRequest.java
@@ -6,6 +6,7 @@ import lombok.Value;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Value
@@ -13,7 +14,7 @@ public class MatriculaRequest {
     UUID idCliente;
     UUID idServico;
     TipoPagamento tipoPagamento;
-    Double valorEntrada;
+    BigDecimal valorEntrada;
     int desconto;
     @Min(value = 1, message = "O valor mínimo é 1")
     @Max(value = 12, message = "O valor máximo é 12")
